made some cleaning changes
